DEV: move overwolf service to lib
