+++ libs/shared/feature/overwolf/src/index.ts
@@ -1,1 +1,2 @@
+export * from './lib/overwolf.service';
 export * from './lib/shared-feature-overwolf.module';
