+++ libs/shared/feature/overwolf/src/lib/shared-feature-overwolf.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { OverwolfService } from './overwolf.service';
 
 @NgModule({
-  imports: [CommonModule],
+	imports: [CommonModule],
+	providers: [OverwolfService],
 })
 export class SharedFeatureOverwolfModule {}
